Update README.md
So, what exactly is "📧 Email Interface for Blind peoples using Python and Ai"?Creating an email interface for blind peoples using Python and AI involves using text-to-speech (TTS) and speech-to-text (STT) technologies. The interface created in Python for the Voice Mail System 💻 project that facilitates access to email for people who are visually impaired. Over here, for stimulation purpose the system💻 aims at providing voice command interaction as well as e-mail related activities such as writing, reading, sending and receiving. The system will be able to receive orders from the user, which they will then complete as specified. Also, the user is able to communicate by composing and receiving email messages as per the preferred language.